% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocorrect-package.R
\docType{package}
\name{allocorrect-package}
\alias{allocorrect}
\alias{allocorrect-package}
\title{allocorrect: Allometric Adjustment for Multi-Group Data}
\description{
This package implemens some correction tools for morphometric variables that are allometrically linked to some undderlying confounding variable, such as body size. It attempts at correcting caveats that arise in previous methods when multiple groups are considered and compared, such as in studies of sexual dimorphism in traits that must first be corrected for body size.
}
\author{
\strong{Maintainer}: Raphaël Scherrer \email{r.scherrer@rug.nl} (\href{https://orcid.org/0000-0002-1447-7630}{ORCID})

}
\keyword{internal}
