% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_data.R
\name{correct_data}
\alias{correct_data}
\title{Correct for varying allometries among groups}
\usage{
correct_data(data, transform = NULL, linear = TRUE, separate = TRUE)
}
\arguments{
\item{data}{The data set to use, as a data frame or tibble}

\item{transform}{Optional transformation function used to re-scale the
vertical axis}

\item{linear}{Whether the lines should be straight or allometric}

\item{separate}{Whether the mid-line should be an average of separate, group-specific
lines or the result of a global regression through the whole data set}
}
\value{
An updated tibble where the variable of interest has been corrected.
}
\description{
This function performs allometry correction on some data when the allometric
relationship may differ among groups.
}
\details{
The function first applies the requested transformation, fits a
regression model (either linear, \code{y ~ x}, or allometric, \code{log10(y)
 ~ log10(x)}) to come up with a mid-line (either through global regression
or averaging group-specific regressions --- which is the recommended default),
and computes the corrected values of the variable of interest as deviations
from that mid-line.
}
\examples{

# Simulate some data
data <- simulate_data(
  intercept_f = 0.1, 
  intercept_m = 10, 
  slope_f = 0.5, 
  slope_m = 0.5,
  n = 20L
)

# Correct for allometry
correct_data(data)

}
\seealso{
\code{plot_data}
}
