% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Extract predictions from a fitted model}
\usage{
get_predictions(fit, data, linear, separate = FALSE)
}
\arguments{
\item{fit}{The fitted model (by \code{lm()})}

\item{data}{The data frame to \code{predict} values for}

\item{linear}{Whether the model is linear (if not, an allometric relationship
is assumed and the inverse of \code{log10} will be applied to the predictions)}

\item{separate}{Whether to compute the average of separate group-specific
predictions or not (i.e. global regression)}
}
\value{
A vector of predicted values for the variable that the fitted model
\code{fit} is supposed to model.
}
\description{
This function extracts predicted values for a given data set from a fitted
model, either fitted separately through multiple groups (i.e. with an
interaction term) or a global regression through the whole data set.
}
\details{
This function is basically a wrapper around \code{predict} and can
work on a model, fitted with \code{lm} and regressing a certain variable
against another, either linearly (\code{y ~ x}) or allometrically (
\code{log10(y) ~ log10(x)}), and on a group-specific basis or not (i.e.
with one regression line per group if \code{separate} is \code{TRUE}).
See \code{?predict} for more details.
}
\examples{

# Simulate some data
data <- simulate_data(
  intercept_f = 0.1, 
  intercept_m = 10, 
  slope_f = 0.5, 
  slope_m = 0.5,
  n = 20L
)

# Fit a linear model
fit <- lm(y ~ x * sex, data)

# Extract predictions
get_predictions(fit, data, linear = TRUE, separate = TRUE)

}
\seealso{
\code{correct_data}, \code{plot_data}
}
