% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate allometric data}
\usage{
simulate_data(
  intercept_f,
  intercept_m,
  slope_f,
  slope_m,
  n = 1000L,
  mean_x = 25,
  sd_x = 5,
  sd_y = 5,
  fun = function(x) x,
  homogeneity = 1
)
}
\arguments{
\item{intercept_f, intercept_m, slope_f, slope_m}{Linear model parameters for both sexes}

\item{n}{Number of data points per sex}

\item{mean_x}{Mean value along the horizontal axis}

\item{sd_x}{Standard deviation along the horizontal axis}

\item{sd_y}{Noise to add to the (normally distributed) data}

\item{fun}{Optional function to convert the relationship into a non-linear allometry}

\item{homogeneity}{Degree of range cover of both groups along the horizontal axis (0 to 1).
A value of 1 means that the full range will be covered by both sexes on the
horizontal axis, and the closer to 0, the more each sex will be restricted
to one side of the range (by default, females to the left and males to the
right).}
}
\value{
A tibble containing the columns \code{x}, \code{y} and \code{sex}.
}
\description{
This function can be used to simulate a dummy data set with sexually dimorphic
allometry, on which to test the correction tools of the package.
}
\details{
The function simulates a variable \code{x} that depends on some
underlying variable \code{y} according to some linear relationship that can differ
between males and females, and then transforms those data so the relationship
is no longer linear (if requested). Noise is added to the simulated data before
transformation.
}
\examples{

# Simulate some data
simulate_data(
  intercept_f = 0.1, 
  intercept_m = 10, 
  slope_f = 0.5, 
  slope_m = 0.5,
  n = 20L
)

}
\seealso{
\code{plot_data}, \code{correct_data}
}
